pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentprobe-framework"
version = "1.0.0"
description = "A testing and evaluation framework for software agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "AgentProbe Contributors" },
]
keywords = ["testing", "agents", "evaluation", "trace", "software-testing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.5,<3",
    "click>=8.0,<9",
    "pyyaml>=6.0,<7",
    "aiohttp>=3.9,<4",
    "rich>=13.0,<14",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.29,<1",
]
dashboard = [
    "fastapi>=0.109,<1",
    "uvicorn>=0.27,<1",
]
eval = [
    "numpy>=1.26,<2",
]
dev = [
    "ruff>=0.3",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
    "pre-commit>=3.6",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "coverage[toml]>=7.4",
    "httpx>=0.27",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
agentprobe = "agentprobe.cli:main"

[project.entry-points."pytest11"]
agentprobe = "agentprobe.pytest_plugin"

[project.urls]
Homepage = "https://github.com/dyrach1o/agentprobe-framework"
Documentation = "https://dyrach1o.github.io/agentprobe-framework"
Repository = "https://github.com/dyrach1o/agentprobe-framework"
Issues = "https://github.com/dyrach1o/agentprobe-framework/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agentprobe"]

# ── Ruff ──

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
    "PTH",   # flake8-use-pathlib
    "ERA",   # commented-out code detection
    "PL",    # pylint subset
    "PERF",  # performance anti-patterns
]
ignore = [
    "E501",    # line length (handled by formatter)
    "PLR0913", # too many arguments (common in config objects)
    "TC001",   # TYPE_CHECKING import moves — all files use `from __future__ import annotations`
    "TC003",   # TYPE_CHECKING import moves — all files use `from __future__ import annotations`
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLR2004",  # magic value comparisons are normal in tests
    "PLC0415",  # imports inside functions are OK in tests
]
"src/agentprobe/eval/*.py" = [
    "PLC0415",  # lazy imports of `os` to avoid import overhead on module load
]
"src/agentprobe/cli/commands/dashboard.py" = [
    "PLC0415",  # lazy imports of fastapi/uvicorn for optional dependency
]
"src/agentprobe/testing.py" = [
    "PLC0415",  # lazy import of TestCase inside assert_score
]
"src/agentprobe/pytest_plugin.py" = [
    "PLC0415",  # lazy import of TestCase inside evaluate
]
"examples/*.py" = [
    "PLC0415",  # lazy imports in examples for optional dependencies
    "PLR2004",  # magic values are fine in examples
]

[tool.ruff.lint.isort]
known-first-party = ["agentprobe"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ── Mypy ──

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "fastapi.*",
    "uvicorn.*",
    "httpx.*",
]
ignore_missing_imports = true

# ── Pytest ──

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=agentprobe",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
    "api: marks tests that make real API calls (expensive)",
    "e2e: marks end-to-end tests with real frameworks",
    "agentprobe: marks tests using the AgentProbe framework",
]

# ── Coverage ──

[tool.coverage.run]
source = ["agentprobe"]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@overload",
]
